chore: alteração de variáveis de valores de double para BigDecimal.

+++ AtividadeFinal/Projeto/Pedido.java
@@ -1,14 +1,15 @@
 package Projeto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class Pedido {
     private int id;
     private Cliente cliente;
     private List<Pizza> pizzas;
-    private double valorTotal;
+    private BigDecimal valorTotal;
 
-    public Pedido(int id, Cliente cliente, List<Pizza> pizzas, double valorTotal){
+    public Pedido(int id, Cliente cliente, List<Pizza> pizzas, BigDecimal valorTotal){
         this.id = id;
         this.cliente = cliente;
         this.pizzas = pizzas;
